pyproject.toml
[project]
authors = [
    {name = "Padraic Slattery", email = "pgoslatara@gmail.com"}
]
dependencies = [
    "click>=8.2.1",
    "google-adk>=1",
    "o365>=2.1.2"
]
description = "Add your description here"
license-files = ["LICENSE"]
name = "name"
readme = "README.md"
requires-python = ">=3.12,<3.13"
version = "0.0.0"

[project.optional-dependencies]
dev = [
    "ipykernel",
    "mypy~=1.0",
    "pre-commit~=4.0",
    "ruff~=0.0",
    "types-requests"
]

[project.scripts]
misstea = "misstea.main:cli"
outlook_token = "misstea.sub_agents.outlook.utils:generate_outlook_token"

[tool.black]
line-length = 99
include = '/project$|\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.eggs
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
follow_imports = "silent"
no_implicit_reexport = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
ignore = ["COM812", "D100", "D203", "D213", "D406", "D407", "D409", "E501", "G004", "S101"]
preview=true
select = ["A", "ARG", "B", "C4", "COM", "D", "DOC", "E", "EXE", "F", "I", "ICN", "LOG", "G", "N", "PT", "PTH", "Q", "RUF", "S", "SIM", "T20", "TCH", "TID", "W"]
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"tests/*" = ["D100", "D103", "D104", "D400", "D415", "E402"]

[tool.ruff.format]
docstring-code-format = false
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.uv]
package = true
